+++ src/courses/courses.module.ts
@@ -5,6 +5,8 @@ import { CategoriesController } from './categories.controller';
 import { CategoriesService } from './categories.service';
 import { CoursesController } from './courses.controller';
 import { CoursesService } from './courses.service';
+import { CurriculumController } from './curriculum.controller';
+import { CurriculumService } from './curriculum.service';
 import { Category } from './entities/category.entity';
 import { CourseFaq } from './entities/course-faq.entity';
 import { CourseModule } from './entities/course-module.entity';
@@ -17,8 +19,8 @@ import { Lesson } from './entities/lesson.entity';
     TypeOrmModule.forFeature([Category, Course, CourseModule, Lesson, LessonMaterial, CourseFaq]),
     UsersModule,
   ],
-  controllers: [CategoriesController, CoursesController],
-  providers: [CategoriesService, CoursesService],
+  controllers: [CategoriesController, CoursesController, CurriculumController],
+  providers: [CategoriesService, CoursesService, CurriculumService],
   exports: [CoursesService, CategoriesService, TypeOrmModule],
 })
 export class CoursesModule {}